+++ R/lint.R
@@ -389,8 +389,14 @@ reorder_lints <- function(lints) {
 Lint <- function(filename, line_number = 1L, column_number = 1L, # nolint: object_name.
                  type = c("style", "warning", "error"),
                  message = "", line = "", ranges = NULL) {
+  if (length(message) != 1L || !is.character(message)) {
+    cli_abort("{.arg message} must be a character string")
+  }
+  if (is.object(message)) {
+    cli_abort("{.arg message} must be a simple string, but has class {.cls {class(message)}}")
+  }
   if (length(line) != 1L || !is.character(line)) {
-    cli_abort("{.arg line} must be a string.", call. = FALSE)
+    cli_abort("{.arg line} must be a character string.")
   }
   max_col <- max(nchar(line) + 1L, 1L, na.rm = TRUE)
   if (!is_number(column_number) || column_number < 0L || column_number > max_col) {